fix to ix which is deprecated
the .ix formulation in pandas is deprecated and thus it has to be changed into .iloc or .loc. In this commit, I've solved this issue, in order to avoid warning messages at this moment and to avoid error messages in the future, when .ix is not allowed anymore in the future.